Allow environment variables to be specified.
This uses `-e` / `--env` on the command line. One can either specify a
`key=value` or just a `key`. In the latter case, the value is inherited
from the calling environment. If the variable is missing from the
environment nothing is passed. This mirrors the semantics of `docker
run`.

Obviously in the case of the `run` command, the entire environment is
inherited anyway (for now at least), but this will be more useful in the
client-server case.