Add a database to track tasks.
The database has a single table, mapping a task to its status. Tasks are
initially marked as Pending, get updated to Running, and then either
Complete or Failed.

The database uses SQLite and is currently only kept in memory. Obviously
we will need to persist that on disk instead. We will definitely keep
using SQLite for local uses of marathon, but in server use cases we
might end up adding support for PostgreSQL.

The server gains new endpoints for listing all tasks and for checking
the status of a single task. The client has matching commands that
interact with the server.

At the moment the client does not have any local database. Tasks run
locally (with `marathon task run`) are not recorded anywhere.
Similarly, tasks run remotely (with `marathon task submit`) are only
tracked by the server, not locally. This will change in the future.