Introduce a work queue and separate worker processes.
The worker now runs outside of the server process and communicates with
it over HTTP. There is one endpoint used to pull tasks from a work queue
and one endpoint to update the task's status.

The task queue is recorded in the SQLite database as its own table.
While it would be possible to maintain the queue as part of the main
task table, I think the separation of concern is good to have. In
particular the task table will be synchronized across the client and the
server, whereas the work queue does not need to be (and probably should
not).

It would have also been possible to maintain the queue in memory,
although we would then need to worry about maintaining atomicity when
adding or updating new tasks. Using the database is just simpler as it
allows us to use transactions.

As a convenience, the server command has a `--start-worker` flag that
will run a worker thread concurrently. This is useful in local
development and removes the need to have yet another console open.